test: update circuitDir structure
