+++ test/cli/snarkit.ts
@@ -1,16 +1,21 @@
-import {parepareCircuitDir, testCircuitDir} from "../tester/c";
+import { parepareCircuitDir, testCircuitDir } from '../tester/c';
+import * as path from 'path';
+import * as fs from 'fs';
 
 // TODO: use yargs here
 
 async function main() {
-    const command = process.argv[2];
-    if (command == 'compile') {
-        await parepareCircuitDir(process.argv[3], false);
-    } else if (command == 'test') {
-        await testCircuitDir(process.argv[3], '');
-    } else {
-        throw new Error('invalid argv ' + process.argv);
-    }
+  const command = process.argv[2];
+  if (command == 'compile') {
+    await parepareCircuitDir(process.argv[3], { alwaysRecompile: false, verbose: true });
+  } else if (command == 'test') {
+    const circuitDir = process.argv[3];
+    await testCircuitDir(circuitDir);
+  } else {
+    throw new Error('invalid argv ' + process.argv);
+  }
 }
 
-main();
+if (require.main === module) {
+  main();
+}
