+++ test/transfer.ts
@@ -5,6 +5,7 @@ import { Account } from '../helper.ts/account';
 import { hashAccountState, calculateGenesisOrderRoot } from '../helper.ts/state-utils';
 import { SimpleTest, TestComponent } from './interface';
 import * as common from './common';
+import { GlobalState } from './global_state';
 
 // circuit-level definitions
 const orderLevels = 2;
@@ -14,7 +15,7 @@ const accountLevels = 5;
 const genesisOrderRoot = calculateGenesisOrderRoot(orderLevels);
 
 function initTestCase() {
-  let state = new common.GlobalState(balanceLevels, orderLevels, accountLevels, 1);
+  let state = new GlobalState(balanceLevels, orderLevels, accountLevels, 1);
 
   const tokenID = 2n;
   const amount = 300n;
